docs: the encoders docstring says fp32, which is what is shipped
The exports were int8 when the module was written and are fp32 now: at fp32
the ONNX path reproduces the torch path byte-for-byte across the golden
snapshots, and int8 does not. The docstring was the last place still naming
the rejected variant.

+++ src/restruct/encoders.py
@@ -9,7 +9,7 @@ it. `onnxruntime` + `tokenizers` + `numpy` is about 40 MB and has no GPU
 variant to resolve into by accident.
 
 The weights are the `model.onnx` written into each model directory by
-`tools/export_onnx.py`, quantized to int8. Tokenization is the same
+`tools/export_onnx.py`, quantized to fp32. Tokenization is the same
 `tokenizer.json` the torch path used, read directly by `tokenizers`.
 
 Two behaviours are reimplemented here because they used to come from a
